requirements.txt
faker
cassandra-driver
